Rename and refactor fyneextensions_test.go to fyneextensions_demo.go in example directory
The renamed file has been moved to the 'example' directory and also refactored. The main aim of this commit is to change the test file into a practical demonstration, showing the usage of fyneextensions. The functions and modules have been updated, specifically modifying 'package fyneextensions' to 'package main', demonstrating the usage of fyneextensions in a standard main package.